cambio en tipo de dato (importante)

+++ ProgramaApuestasJDBC/src/interfaces/Ingreso.java
@@ -5,7 +5,7 @@ public interface Ingreso {
 
     public int getId() ;
     public void setId(int id) ;
-    public int getCantidad();
+    public double getCantidad();
     public void setCantidad(int cantidad);
     public String getDescripcion() ;
     public void setDescripcion(String descripcion) ;
